move duplicate code into PsoXja/globals.lua
